refactor: move mocks folder to internal folder
